requirements.txt
streamlit>=1.28.0
tensorflow>=2.12.0
transformers>=4.21.0
pillow>=9.0.0
numpy>=1.21.0